Beginning to improve security
Have added the 'grabValue()' function and 'parseReference()' functions
to allow toffle to avoid using 'eval()' in that references made using
toffle components can eventually be interpreted by toffle itself and not
just passed to the javascript interpreter. This is a lot safer and means
that we can confine the scope of these references to be whichever
object/s we want (most likely a single object attached to an executing
context that contains just data objects that we need ). It has not been
implemented yet, i may add an option to the initializing object of a
toffle template to specify whether to process the template in a
'sandbox' mode (toffle interprets all toffle reference components. not
as powerful but much safer)