Added Match Statement!
Added a Match statement. Works like each but only uses items that match
on a given helper function. Fixed a bug where an 'each' context would be
processed even if there were no iterations to be made.

Found a number of bugs to be addressed.
- Need to use comma seperation and parenthesis for helper arguments
(just like in template reference arguments) as characters they contain
can mess up parsing the arguments.

Any identifiers that have illegal characters (that would mess up token
parsing) will have to be wrapped in <!^ and ^!>. Every character in this
should be ignored and treated as one big string during compilation.